Added weighted motif discovery
Rewrote Sequences object to contain a weight object, which is for
relative importance of the sequences. This can come from signal data
or concentration data. This required adding a factory method for
Sequences and I fixed some bugs in along the way. Added weighting to
the motif discovery methods as well. I rewrote the relevant generics
methods so that when sequences are indexed or bound, the weights are
retained. I had to reupload the sequence data for example data. I
think I got the same datasets, the AMPs were aligned. I updated the
documentation and tutorial for the new weighting methods. I also
noticed there was a mistake in the SI for White2012Simplifying. The
E-step should have products, not sums. Anyway, I've noted the relevant
code in case anyone ever tries to follow the implementation of that
equation.